Refatoração program
