+++ src/BackEnd/Api/Program.cs
@@ -1,3 +1,4 @@
+using Api.Configurations;
 using Api.Models;
 using Business;
 using Data;
@@ -11,75 +12,11 @@ using System.Text;
 using System.Text.Json.Serialization;
 
 var builder = WebApplication.CreateBuilder(args);
-
-// Add services to the container.
 builder.AddApplicationServices();
 builder.AddDataServices();
 builder.Services.Configure<ApiBehaviorOptions>(options => options.SuppressModelStateInvalidFilter = false);
-//builder.Services
-//    .AddFluentValidationAutoValidation()
-//    .AddFluentValidationClientsideAdapters();
-
 builder.Services.AddControllers();
-	//.AddJsonOptions(options =>
-	//{
-	//	options.JsonSerializerOptions.ReferenceHandler = ReferenceHandler.Preserve;
-	//});
-
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen(c =>
-{
-    c.SwaggerDoc("v1", new OpenApiInfo { Title = "Online store", Version = "v1", });
-    //configuracao para o Swagger utilizar o jwt token
-    c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme()
-    {
-        Name = "Authorization",
-        Type = SecuritySchemeType.ApiKey,
-        Scheme = "Bearer",
-        BearerFormat = "JWT",
-        In = ParameterLocation.Header,
-        Description = "JWT Authorization header using the Bearer scheme. \n\r\r\n Enter 'Bearer' [space] and then your token in the text input below.\n\r\r\nExample: \"Bearer 12345abcdef\"",
-    });
-
-    c.AddSecurityRequirement(new OpenApiSecurityRequirement
-    {
-        { new OpenApiSecurityScheme
-            {
-                Reference = new OpenApiReference
-                {
-                    Type = ReferenceType.SecurityScheme,
-                    Id = "Bearer"
-                }
-            },
-            new string[] {}
-        }
-    });
-});
-
-var jwtSettingSection = builder.Configuration.GetSection("JwtSettings");
-builder.Services.Configure<JwtSettings>(jwtSettingSection);
-var jwtSettings = jwtSettingSection.Get<JwtSettings>();
-var key = Encoding.ASCII.GetBytes(jwtSettings.Secret);
-
-//Configuracao Jwt token
-builder.Services.AddAuthentication(options =>
-{
-    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-}).AddJwtBearer(options =>
-{
-    options.RequireHttpsMetadata = true;
-    options.SaveToken = true;
-    options.TokenValidationParameters = new TokenValidationParameters
-    {
-        IssuerSigningKey = new SymmetricSecurityKey(key),
-        ValidateIssuer = true,
-        ValidIssuer = jwtSettings.Issuer,
-        ValidateAudience = true,
-        ValidAudience = jwtSettings.Audience,
-    };
-});
+builder.Services.AdicionaSwaggerConfiguracao(builder.Configuration);
 
 var app = builder.Build();
 app.UseDbMigrationHelper();
